Refactor: ResponseCode Exception에서 보유하도록 변경 및 클래스명에 규칙 제거
