--- module-api/src/main/kotlin/finn/exception/ExceptionResponseCodeMapper.kt
@@ -1,12 +1,0 @@
-package finn.exception
-
-import finn.response.ResponseCode
-
-fun mapResponseCode(e: Exception): ResponseCode {
-    val className = e::class.simpleName ?: ResponseCode.SERVER_ERROR.toString()
-
-    return ResponseCode.entries.firstOrNull { code ->
-        val keyword = code.name.replace("_", "") // 언더스코어 표기 무시
-        className.contains(keyword, true)
-    } ?: ResponseCode.SERVER_ERROR
-}
