+++ module-domain/src/main/kotlin/finn/exception/DomainPolicyViolationException.kt
@@ -1,9 +1,11 @@
 package finn.exception
 
-class BadRequestDomainPolicyViolationException(
+class DomainPolicyViolationException(
     message: String,
     cause: Throwable?
 ) : CommonException(message, cause) {
+    override val code: ResponseCode = ResponseCode.BAD_REQUEST
+
     constructor(message: String) : this(
         message, null
     )
