+++ module-domain/src/main/kotlin/finn/validator/SearchKeywordMatcher.kt
@@ -1,7 +1,7 @@
 package finn.validator
 
 import finn.entity.Ticker
-import finn.exception.BadRequestDomainPolicyViolationException
+import finn.exception.DomainPolicyViolationException
 import java.util.*
 
 fun checkTickerMatchCondition(ticker: Ticker, keyword: String): Boolean {
@@ -12,6 +12,6 @@ fun checkTickerMatchCondition(ticker: Ticker, keyword: String): Boolean {
 
 fun checkKeywordValid(keyword: String?) {
     if (keyword.isNullOrBlank() || keyword.length < 2) {
-        throw BadRequestDomainPolicyViolationException("키워드는 2글자 이상만 요청할 수 있습니다.")
+        throw DomainPolicyViolationException("키워드는 2글자 이상만 요청할 수 있습니다.")
     }
 }
