+++ module-persistence/src/main/kotlin/finn/exception/CriticalDataPollutedException.kt
@@ -3,7 +3,9 @@ package finn.exception
 /*
     지원하지 않는 옵션이 유입되어 오염
  */
-class ServerErrorCriticalDataPollutedException(message: String, cause: Throwable?) :
+class CriticalDataPollutedException(message: String, cause: Throwable?) :
     CommonException(message, cause) {
+    override val code: ResponseCode = ResponseCode.SERVER_ERROR
+
     constructor(message: String) : this(message, null)
 }
