+++ module-persistence/src/main/kotlin/finn/repository/impl/NewsRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package finn.repository.impl
 
 import finn.entity.News
-import finn.exception.ServerErrorCriticalDataPollutedException
+import finn.exception.CriticalDataPollutedException
 import finn.mapper.toDomain
 import finn.paging.PageResponse
 import finn.queryDto.NewsDataQueryDto
@@ -31,7 +31,7 @@ class NewsRepositoryImpl(
 
             "negative" -> newsQueryRepository.findAllNegativeNewsList(page, size)
 
-            else -> throw ServerErrorCriticalDataPollutedException("filter: $filter, 지원하지 않는 옵션입니다.")
+            else -> throw CriticalDataPollutedException("filter: $filter, 지원하지 않는 옵션입니다.")
         }
         return PageResponse(newsExposedList.content.map { it ->
             toDomain(it)
